Cache User table in translib
Prior to this change, applications needing to get authorization
information for authenticated users would have to resort to fetching the
Linux groups via the HAM NSS modules. This change adds support to
Translib to cache the (new) User table in memory, so that the
applications can simply call the `GetUser` method, and that will either
retrieve it from the cache, or get it from the DB and cache it.

This change also adds support to listen to Redis events so that when the
key is either created, modified or deleted, the cache is updated
accordingly.