[JIRA SONIC-18317] Must expression for delete operation
CVL sematic check step will now remove the delete fields from the
data cache before evaluating the must expressions. This allows to
1) Skip must expressions definied for the delete field.
2) Catch dependency on delete filds due to must expressions defined
for other fields.

Exmple:
    leaf XX {
    }
    leaf YY {
        must "current() > ../XX";
    }

Here, the must condition is not evaluated during YY deletion. However
XX deletion fails if YY is defined (because XX is a hard dependency
for YY). To make Xx optional, the must expression should be modified
as "not(../XX) or current() > ../XX)".

Change-Id: I67bbb00354b7bc568f4c331c350a44e2e24121bb