[JIRA SONIC-22629] VxLAN VtepName restricted to 10 characters
Linux kernel does not create the vxlan netdevice if the name exceeds 10 character limit.
Fix the sonic_vxlan.yang to have a maximum length of only 10 characters for the vtep name.

This will ensure that both the Klish and OC are restricted with the same limit of characters.

Test Case(s)
Using Klish CLI, ensured that the vtepname with more than 10 characters are not accepted in the
klish CLI.
Tested all the vxlan config/show commands along with remote BGP evpn peers with ecmp underlay and
unicast/BUM traffic to ensure that traffic is restored with the klish commands.

Change-Id: Icd6d7cde378a9e27da4f6582ac143c7dbcc17b4c