[JIRA SONIC-23448] Fix the oc-yang rest of the errors
Moved the vxlan-tunnel-info & vxlan-vni-peer-info elements into state container
Changed annotation file to reflect the new URI
Fixed the tunnel status issue to convert into the corresponding field xformer.
With the above changes, all the yang compliant errors are resolved.
This will ensure that the mgmt framework (when all yang validation are enabled), will compile and
link successfully and will not be blocker from vxlan yang perspective.

However, the vxlan-tunnel-info and vxlan-vni-peer-info are functionally not working and needs to be
debugged further.

Change-Id: I8af918af38fa44dc1938c00cd0e0f16aafae736d