[JIRA SONIC-24166] Adding "Ethernet 12" as a portchannel member throws "% Error: Port group member
Both port-channel and speed change shares same transformer function.
The port group membership check is supposed to be done specific to speed change; but it is done in common place.

sonic(config)# interface Ethernet 0
sonic(conf-if-Ethernet0)# speed 10000
%Error: Port group member. Please use port group command to change the speed
sonic(conf-if-Ethernet0)# no speed
% Error: Port group member. Please use port group command to change the speed
sonic(conf-if-Ethernet0)# channel-group 1
sonic(conf-if-Ethernet0)# no channel-group
sonic(conf-if-Ethernet0)# channel-group 1
sonic(conf-if-Ethernet0)#

Signed-off-by: Prasanth Kunjum Veettil <prasanth.kunjumveettil@broadcom.com>
Change-Id: If784640596e27d5034c3cd2e8f786c1f3eed5c44