[JIRA SONIC-18836] Handle NULL,NULL field,value pair
Latest changes in mgmt-framework introduces NULL,NULL as field,value pair when the last field is deleted.
For storm-control, when "kbps" field DELETE is triggered via CURL, NULL,NULL is introduced for the given key (IntfName|storm-type)
When user executes show command "show storm-control", it displays empty column for the row which has NULL,NULL.

Fix is to handle "kbps" field delete by deleting the key instead of deleting just the field/value pair.
Change is done in intf_post_xfmr as only one post_xfmr can be registered for a given module (openconfig-interface in this case).

UT - Verified (on the merge branch) that the key gets deleted (instead of adding NULL,NULL) when "kbps" delete is triggered

Change-Id: I1e4bbf9a3c60f6163d91f6da4cd32e04e82d2c03