[JIRA SONIC-24821] - Fix issue in error reporting
If a deviate statement uses type which is a typedef,
instead of type statement, the error is pointing to a
typedef statement and if the same typedef is used in
multiple deviate statements, an only a single error is reported.
pointing to the typedef statement is not right
because the typedef can be used in a non-deviate statement also.
it's better to point to deviate stmt.

Change-Id: I99e02345693ca2dd2892099148f2e2a1aac866a6