[JIRA SONIC-25427] IGMP Snooping entry not deleted on disabling via REST PATCH request
Fix is to delete the parent node when igmp snooping "enabled" field is set to "false".

UT
enabled igmp snooping on VLAN and added additional parameters (querier,fast-leave,query-interval,...)
initiated a PATCH request for "enabled" field to set to "false" and observed that the entry itself is deleted.
PATCH request to set to "true" and observed that the entry is created.

TODO
When DELETE req is done for the last field of CFG_L2MC_TABLE, NULL,NULL field is inserted. This needs to be fixed.

Change-Id: I7a5897594c4a33e64f6794aecc74662cc065a4d6