[JIRA SONIC-25170] source vtep ip should not be mandatory argument
Retrieve the input values of sip, qos-mode and dscp even when the vxlanIntfName value is empty.
This would ensure that the interface gets created with the parameters input by user.

UT
Checked that the VTEP creation succeeds with any or all inputs for sip/qos-mode/dscp present in the REST POST request.

Change-Id: I5f580805342ef9ce07efd6d1612bf06897e95e95