Connect to the DB only when retrieving the user info
Prior to this change, userDb was a persistent connection to the
ConfigDB, which was unused for the majority of the time. The only place
where we need a connection is when we need to retrieve the user
information, and therefore we can move the DB create, and eventual
delete to the getUserFromDB function.