[JIRA SONIC-30428] Change default interface command to not change a port-group member port speed
The "default interface" command is modifying the speed of a port group member port.
This is causing the statistics to be cleared. Which in turn results in a momentary
state where invalid counter values are returned.

There are two problems here:
1. default interface command should not modify port group member speed as the
port-group member's speed is not its individual property but is of the port-group
as a whole.

2. When port-group speed is changed, the stats are going awry for a few seconds.

The proposed changes are to change the speed of a port-group member to valid speeds
when reverting the port to its default configuration state. This addresse
only issue#1

Change-Id: I867074fb91547e4ace298b50069fccc580814c55