[JIRA SONIC-30082] Fixed VRF_VNI OC Yang URI Restconf issue
since VRF_VNI mapping has different key from VLAN_VNI and vni mapping of VRF
is in the VRF table unlike VLAN VNI keys are different between both. So
isVirtuablTbl is required for the VRF VNI mapping in subscriber
Because of this rest conf compliance for VRF_VNI case was not working
So fixed it.

Change-Id: Idf67a68dc1b4f6b45e35747ef6e71a5a0161cca6