[JIRA SONIC-31503] - SAG IP is not cleaned up when VLAN is deleted
Upon SAG IP delete, vrrporch fails to find the vrf associated with this RIF causing
SAG IP to be retained in ASIC_DB. Fix for SONIC-31015 has allowed deletion of VLAN
when SAG IP is configured. This has created ordering issues in OA and sometimes cleanup
is not done properly.
Reverting fix done for SONIC-31015

Change-Id: I1f5dc3ec1d3922f7fe59881af45d0466791c5486