[JIRA SONIC-33171] Fix restoring fec mode to its default value using the default interface command
Deleting fec configuration field in the CONFIG_DB is not setting fec mode to none in APPL_DB.
If the default fec mode is not defined in the factory default configuration file,
the PORT table will not have a fec field.

However, once a valid fec mode is configured by the user, it is expected that
the fec field is to "none" to turn off the fec mode.

The proposed changes to the default interface command performs the following action:
Condition-1: Check if a valid fec mode is configured on the port
Condition-2: Check if a valid fec mode is defined in the default configuration file
fec is set to "none" if the Condition-1 is true and the Condition-2 is false.

Change-Id: Id9b8b8e1a7c3051caba889d58c9ce7bfc8820324