[JIRA SONIC-34096] Translib subscribe infra fixes
1) Use target node's parent path as 'prefix' for initial sync updates
and when target path is a leaf.

2) Treat DB entry/field deletes as yang deletes. We may have to
introdue a flag in notificationAppInfo for apps to override this
behavior.

3) Automatically disable on-change for non-DB fields in translib infra

4) Refactor translateSubscribe and processSubscribe APIs
Introduced request and response structures for translateSubscribe and
processSubscribe APIs. This allows future extensions to these APIs
without having to change app module function signatures.

5) Pass a contextID to app translateSubscribe and processSubscribe
functions. Apps can use it for logging.

6) Fixed notificationInfoBuilder to handle leaf path target by filling
empty relative path & leaf name in db-to-yangpath map.

Change-Id: I1aeca473d14b988a395cf5186c671551ba30ceb8