[JIRA SONIC-37018] Avoid duplicate init sync messages
When a yang leaf-list is mapped to its own db table, the initial sync
logic should group the db keys based on parent key components. Apps can
indicate the group key components in processSubscribe callback. A new
field processSubResponse.keyGroupComps is introduced, which is a list of
key component indicies for grouping the keys (of leaf-list instances).
Transformer based apps should implement a path transformer to handle
such cases. A new field XfmrDbToYgPathParams.keyGroup should be filled
by path transformer callbacks to indicate the key component indices for
the key group. This is an optional field. Apps need not touch it for
other cases.

Change-Id: Ib026a30a0beddc16089dcabca555ed0b5234e230