[JIRA SONIC-37099] Treat empty ygot struct as delete notification
1) Transformer returns empty ygot struct (all fields nil) for Get of
non-existent container, leaf-list and list paths. Subscribe infra
treats such empty struct as 'not found' case and will send a delete
notification for the path. Earlier it used to result in an update with
no values, which does not convey actual context always.

2) Ignore @ suffix from OnChangeCacheDiff field while comparing with
db-to-yang field mappings in notificationInfo. Apps are not suffixing
leaf-list fields with @.

Change-Id: Ic357819bc623bcfb8975a43495a94b9078ecdf67