[JIRA SONIC-39846] Fix setting default configuration of fec on a broken out interface
When using the default interface command on a broken out interface,
the default FEC mode supported on the interface needs to be obtained
based on the platform supported value based on its speed and number of
lanes. Use appropriate API to query the default fec mode for a port
and use it.

Change-Id: Ic3554a7d73f94cf9492a18a12112937ecc4487a0