[JIRA SONIC-41834] Pass updated/deleted db entry to path trasformer
In some cases path transformer cannot resolve all yang path keys from
the db key alone; it needs to load the db entry. But this will not
work for delete case as db entry would have been already deleted. Now
subscribe infra will pass the db entry also to the processSubscribe
callback; which eventually passes the same to path transformer. For
entry create and update cases the value passed will be the latest
value read from db. This does not affect performance as the infra
anyway reads the db value to calculate the diff. Same value will be
passed to the path transformer. For entry delete case, it will be the
value from the subscribe cache.

Path transformer must not modify the fields of the db entry it
receives. It may affect subsequent notifications.

Change-Id: If7933c99d3f87464ada5a1c3b02e8b71941405ae